refactor filters handling for reusabilty
